Kruskal's Algorithm
A java implementation of Kruskal's Algorithm for finding a Minimum
Spanning Tree. The program generates a user-specified number of random
points in 2-, 3-, or 4-dimensional space, and then calculates the weight
of the minimum tree spanning it. There is also the option to generate
random weights for edges in 2-d space (rather than using the euclidean
distance).